refactor main GUI code to video_gui.cpp
