+++ video/ui/main_settings_window.h
@@ -2,13 +2,7 @@
 
 #include <vector>
 
-#include <opencv2/opencv.hpp>
-
-#define CVUI_DISABLE_COMPILATION_NOTICES
-#include "cvui.h"
-
-// EnhancedWindow.h must be included after cvui.h
-#include "EnhancedWindow.h"
+#include "protected_cvui.h"
 
 #include "../device/input_device.h"
