+++ video/ui/main_window.h
@@ -1,12 +1,8 @@
 #pragma once
 
-#include <memory>
+#include <vector>
 
-#define CVUI_DISABLE_COMPILATION_NOTICES
-#include "cvui.h"
-
-// EnhancedWindow.h must be included after cvui.h
-#include "EnhancedWindow.h"
+#include "protected_cvui.h"
 
 #include "../device/input_device.h"
 #include "../opencv/opencv_window.h"
